app/requirements.txt
pymysql            # 추가
streamlit
sentence-transformers
openai
transformers
torch
pandas
numpy
python-dotenv
folium
sqlalchemy
mysql-connector-python
googlemaps
fuzzywuzzy
python-Levenshtein
streamlit-folium   # 추가

